feat: support android (closes #254)
