+++ wolf/Cargo.toml
@@ -10,56 +10,62 @@ repository = "https://github.com/WolfEngine/WolfEngine/"
 version = "3.0.3"
 
 [lib]
-crate-type = ["rlib", "cdylib"]
+crate-type = ["rlib", "cdylib", "staticlib"]
 name = "wolf"
 path = "src/lib.rs"
 
 # before adding any dependency please check it out from
 # https://android.googlesource.com/platform/external/rust/crates/
 
-# Shared dependencies between Wasm & native
+# shared dependencies between all
 [dependencies]
 #brotli = { version = "3.3.3" }
 anyhow = {version = "1.0.56", default-features = false}
-#cfg-if = { version = "1.0.0", default-features = false}
-instant = {version = "0.1.12", default-features = false, features = ["wasm-bindgen"]}
+cfg-if = { version = "1.0.0", default-features = false}
 signals2 = {version = "0.3.2", default-features = false}
-wgpu = {version = "0.12.0", default-features = false, features = ["webgl"]}
 winit = {version = "0.26.1", default-features = false, features = ["x11", "wayland", "wayland-dlopen"]}
 
-# Wasm dependencies
+# dependencies of wasm 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 console_error_panic_hook = { version = "0.1.7", default-features = false }
+instant = {version = "0.1.12", default-features = false, features = ["wasm-bindgen"]}
 rhai = { version = "1.6.0", default-features = false, features = ["sync", "internals", "wasm-bindgen"] }
 serde = { version = "1.0.136", default-features = false, features = ["derive"]}
 wasm-bindgen = { version = "0.2.79", default-features = false }
 wasm-bindgen-futures = { version = "0.4.29", default-features = false }
 wasm-mt = {version = "0.1.2", default-features = false }
 web-sys = {version = "0.3.56", default-features = false, features = ["console", "Document", "Window", "Element"]}
+wgpu = {version = "0.12.0", default-features = false, features = ["webgl"]}
 
-# Native dependencies
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-#scylla = {default-features = false, features = ["ssl"], version = "0.4.3"}
+# shared dependencies between arm, aarch64 & x86_64
+[target.'cfg(any(target_arch = "arm", target_arch = "aarch64", target_arch = "x86_64"))'.dependencies]
+instant = {version = "0.1.12", default-features = false}
+pollster = {version = "0.2.5", default-features = false}
+rhai = { version = "1.6.0", default-features = false, features = ["sync", "internals"]}
+rustls-pemfile = {version = "1.0.0", default-features = false}
+tokio = {version = "1.17.0", default-features = false, features = ["full"]}
+tokio-rustls = {version = "0.23.3", default-features = false, features = ["logging", "tls12"]}
+tonic = {version = "0.7.1", default-features = false, features = ["codegen", "prost", "tls", "compression"]}
+webpki-roots = {version = "0.22.3", default-features = false}
+wgpu = {version = "0.12.0", default-features = false}
+
+# dependencies of x86_64
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+# #scylla = {default-features = false, features = ["ssl"], version = "0.4.3"}
 async-raft = {version = "0.6.1", default-features = false}
 async-trait = {version = "0.1.53", default-features = false}
-futures = {version = "0.3.21", default-features = false}
+futures = { version = "0.3.21", default-features = false }
 hyper = {version = "0.14.18", default-features = false}
 may = {version = "0.3.21", default-features = false}
 memstore = {version = "0.2.0", default-features = false}
 mimalloc = {version = "0.1.28", default-features = false, features = ["secure"]}
 parking_lot = {version = "0.12.0",  default-features = false, features = ["deadlock_detection"]}
-prost = {version = "0.10.0", default-features = false}
-rhai = { version = "1.6.0", features = ["sync", "internals"], default-features = false }
-rustls-pemfile = {version = "0.3.0", default-features = false}
+prost = {version = "0.10.1", default-features = false}
 serde_json = {version = "1.0.79", default-features = false, features = ["std"]}
-sysinfo = {version = "0.23.8", default-features = false, features = ["multithread"]}
-tokio = {version = "1.17.0", default-features = false, features = ["full"]}
-tokio-rustls = {version = "0.23.3", default-features = false, features = ["logging", "tls12"]}
+sysinfo = {version = "0.23.10", default-features = false, features = ["multithread"]}
+tracing = {version = "0.1.34", default-features = false, features = ["std", "attributes"]}
 tokio-tungstenite = {version = "0.17.1", default-features = false, features = ["connect", "rustls"]}
-tonic = {version = "0.7.1", default-features = false, features = ["codegen", "prost", "tls", "compression"]}
-tracing = {version = "0.1.32", default-features = false, features = ["std", "attributes"]}
 uuid = {version = "0.8.2", default-features = false, features = ["v5"]}
-webpki-roots = {version = "0.22.2", default-features = false}
 
 [build-dependencies]
 #bindgen = { version = "0.59.2" }
