+++ wolf/README.md
@@ -18,22 +18,18 @@ This Wolf is a comprehensive set of Rust/C libraries for realtime rendering, rea
 ## Build
 - **Wolf 2/1** via CMake
 - **Wolf 3 via Nightly toolchain**
-  ```bash
-  rustup default nightly
-  ```
   - For **Webassembly** :\
   First you need to setup wasm32
   ```bash
+  rustup default nightly
   rustup target add wasm32-unknown-unknown
-  ```
-  Now run hello wolf demo
-  ```bash
   ./build-wasm.sh
   ./run-wasm.sh
   ```
   Finally the demo will be served at http://localhost:8000
   - For **Native** :
   ```bash
+  rustup default stable
   cd wolf-demo
   cargo run
   ```
