+++ wolf/src/system/mod.rs
@@ -3,9 +3,9 @@ pub mod macros;
 pub mod os;
 pub mod script;
 
-#[cfg(not(feature = "wasm"))]
+#[cfg(not(any(target_os = "android", target_os = "ios", target_arch = "wasm32")))]
 pub mod algorithm;
-#[cfg(not(feature = "wasm"))]
 pub mod net;
+
 // pub mod compression;
 // pub mod db;
