+++ wolf/src/system/net/mod.rs
@@ -1,6 +1,16 @@
 pub mod callback;
+#[cfg(not(target_arch = "wasm32"))]
 pub mod grpc;
-pub mod tcp;
-pub mod tcp_fiber;
+pub mod protocols;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod tcp_client;
+#[cfg(not(any(target_os = "android", target_os = "ios", target_arch = "wasm32")))]
+pub mod tcp_server;
+#[cfg(not(any(target_os = "android", target_os = "ios", target_arch = "wasm32")))]
+pub mod tcp_server_fiber;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod timeouts;
+#[cfg(not(any(target_os = "android", target_os = "ios", target_arch = "wasm32")))]
 pub mod tls;
+#[cfg(not(target_arch = "wasm32"))]
 pub mod udp;
