+++ wolf/src/system/net/protocols.rs
@@ -1,0 +1,7 @@
+pub const MAX_MSG_SIZE: usize = 1024; //1K
+
+#[derive(Debug, Copy, Clone)]
+pub enum TcpProtocol {
+    TcpNative = 0,
+    TcpWebsocket,
+}
