+++ wolf/src/system/net/tls.rs
@@ -1,8 +1,8 @@
 use anyhow::{anyhow, ensure, Result};
-use std::{fs::File, io::BufReader, path::Path, sync::Arc};
+use std::{path::Path, sync::Arc};
 use tokio_rustls::{
-    rustls::{self, Certificate, OwnedTrustAnchor, PrivateKey, RootCertStore, ServerConfig},
-    webpki, TlsAcceptor,
+    rustls::{Certificate, PrivateKey, ServerConfig},
+    TlsAcceptor,
 };
 
 #[derive(Debug)]
@@ -33,40 +33,6 @@ pub fn load_private_keys(p_path: &Path, p_type: &TlsPrivateKeyType) -> Result<Ve
             }
         })
         .map_err(std::convert::Into::into) //e.into()
-}
-
-/// Load root certificate from specific path
-/// # Panics
-///
-/// Will panic if cert is not valid
-pub fn load_root_ca(p_root_ca_path: Option<&Path>) -> Result<RootCertStore> {
-    let mut root_cert_store = rustls::RootCertStore::empty();
-
-    if let Some(ca_path) = p_root_ca_path {
-        let mut pem = BufReader::new(File::open(ca_path)?);
-        let certs = rustls_pemfile::certs(&mut pem)?;
-        let trust_anchors = certs.iter().map(|cert| {
-            let ta = webpki::TrustAnchor::try_from_cert_der(&cert[..]).unwrap();
-            OwnedTrustAnchor::from_subject_spki_name_constraints(
-                ta.subject,
-                ta.spki,
-                ta.name_constraints,
-            )
-        });
-        root_cert_store.add_server_trust_anchors(trust_anchors);
-    } else {
-        root_cert_store.add_server_trust_anchors(webpki_roots::TLS_SERVER_ROOTS.0.iter().map(
-            |ta| {
-                OwnedTrustAnchor::from_subject_spki_name_constraints(
-                    ta.subject,
-                    ta.spki,
-                    ta.name_constraints,
-                )
-            },
-        ));
-    }
-
-    Ok(root_cert_store)
 }
 
 pub fn init_tls_acceptor(
