+++ wolf/src/system/os/mod.rs
@@ -1,6 +1,7 @@
 pub mod w_runtime;
 pub mod w_sigslot;
-#[cfg(not(feature = "wasm"))]
+
+#[cfg(not(any(target_os = "android", target_os = "ios", target_arch = "wasm32")))]
 pub mod w_system;
 
 //#[cfg(target_os = "windows")]
