+++ wolf/src/system/script/mod.rs
@@ -1,4 +1,3 @@
 #[cfg(feature = "wasm")]
 pub mod w_javascript;
-
 pub mod w_rhai;
